requirements.txt
# Core dependencies
flask>=2.3.3
flask-sqlalchemy>=3.1.1
python-dotenv>=1.0.0
langchain>=0.0.267
langchain-community>=0.0.10
langchain-openai>=0.0.2
pydantic>=2.4.2
psutil>=5.9.5

# Document loader dependencies
unstructured>=0.10.30
pypdf>=3.17.0
PyPDF2>=3.0.0
pymupdf>=1.23.0
pdfplumber>=0.10.0
docx2txt>=0.8
python-docx>=0.8.11
beautifulsoup4>=4.12.2
pandas>=2.2.3
openpyxl>=3.1.5
xlrd>=2.0.2
tabulate>=0.9.0  # For better table formatting

# Vector store dependencies
chromadb>=0.4.18
faiss-cpu>=1.7.4
pinecone-client>=2.2.4

# Embedding dependencies
sentence-transformers>=2.2.2
openai>=1.3.0

# LLM dependencies
transformers>=4.35.2
torch>=2.1.1
huggingface_hub>=0.19.4

# Testing dependencies
pytest==7.4.0
pytest-flask==1.2.0

python-magic-bin
markdown